Update rebuild.ts
The fix is needed for controlling the OS during the assembly process. For example, there is currently an issue when using electron-builder version 24, which started using electron/rebuild. The problem can be found here: [link to the specific line of code on GitHub](https://github.com/electron-userland/electron-builder/blob/46524169cefbfa18e342d7fa19e79e710aae848e/packages/app-builder-lib/src/util/yarn.ts#L162).

The OS will always be the current one, and because of this, it is not possible to download the necessary bindings.